A bunch of various adds
The grass material is for level 1, it looks alright over such a large terrain. 
The sound cues are needed to play the foot steps, i turned them down so they don't get ear grating.
The Animation BP has two new events inside of it that play either sound, in which you might to replace the sound because of the github transfer.
The animation run holds the timing in which the sounds are played. Replace yours with the new one.
Meme man has updated max movement speed.